+++ locations/location_mgr.go
@@ -19,9 +19,10 @@ package locations
 import (
 	"context"
 	"encoding/json"
-	"github.com/ystia/yorc/v4/locations/adapter"
 	"path"
 	"time"
+
+	"github.com/ystia/yorc/v4/locations/adapter"
 
 	"github.com/hashicorp/consul/api"
 	"github.com/pkg/errors"
@@ -45,12 +46,14 @@ type LocationConfiguration struct {
 type Manager interface {
 	InitializeLocations(locationFilePath string) (bool, error)
 	CreateLocation(lConfig LocationConfiguration) error
-	RemoveLocation(locationName, locationType string) error
+	UpdateLocation(lConfig LocationConfiguration) error
+	RemoveLocation(locationName string) error
 	SetLocationConfiguration(lConfig LocationConfiguration) error
 	GetLocations() ([]LocationConfiguration, error)
 	GetLocationProperties(locationName, locationType string) (config.DynamicMap, error)
 	GetLocationPropertiesForNode(ctx context.Context, deploymentID, nodeName, locationType string) (config.DynamicMap, error)
 	GetPropertiesForFirstLocationOfType(locationType string) (config.DynamicMap, error)
+
 	Cleanup() error
 }
 
@@ -80,6 +83,20 @@ func GetManager(cfg config.Configuration) (Manager, error) {
 	return locationMgr, nil
 }
 
+// NewManager creates a Location Manager for the http server
+// The http server has already a Consul client instance that was
+// provided to it as a constructor input parameter (see rest.NewServer function).
+func NewManager(client *api.Client) Manager {
+
+	var locationMgr *locationManager
+	if locationMgr == nil {
+		locationMgr = &locationManager{cc: client}
+		locationMgr.hpAdapter = adapter.NewHostsPoolLocationAdapter(client)
+	}
+
+	return locationMgr
+}
+
 // CreateLocation creates a location. Its name must be unique.
 // If a location already exists with this name, an error will be returned.
 func (mgr *locationManager) CreateLocation(lConfig LocationConfiguration) error {
@@ -90,13 +107,26 @@ func (mgr *locationManager) CreateLocation(lConfig LocationConfiguration) error 
 	return mgr.createLocation(lConfig)
 }
 
-// RemoveLocation removes a given location
-func (mgr *locationManager) RemoveLocation(locationName, locationType string) error {
+// UpdateLocation updates a location. Its name must exist.
+// If no location exists with this name, an error will be returned.
+func (mgr *locationManager) UpdateLocation(lConfig LocationConfiguration) error {
 
-	if locationType == adapter.AdaptedLocationType {
+	if lConfig.Type == adapter.AdaptedLocationType {
+		return mgr.hpAdapter.SetLocationConfiguration(lConfig.Name, lConfig.Properties)
+	}
+	return mgr.updateLocation(lConfig)
+}
+
+// RemoveLocation removes a given location
+func (mgr *locationManager) RemoveLocation(locationName string) error {
+	isHostPoolLocationType, err := mgr.checkIsHotPoolLocationType(locationName)
+	if err != nil {
+		return err
+	}
+	if isHostPoolLocationType {
 		return mgr.hpAdapter.RemoveLocation(locationName)
 	}
-	return mgr.removeLocation(locationName, locationType)
+	return mgr.removeLocation(locationName)
 }
 
 // SetLocationConfiguration sets the configuration of a location.
@@ -329,7 +359,28 @@ func (mgr *locationManager) createLocation(configuration LocationConfiguration) 
 		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
 	}
 	if kvp != nil && len(kvp.Value) != 0 {
-		return errors.Errorf("location %q already exists", configuration.Name)
+		return errors.WithStack(locationAlreadyExistError{})
+	}
+
+	return mgr.setLocationConfiguration(configuration)
+}
+
+func (mgr *locationManager) updateLocation(configuration LocationConfiguration) error {
+
+	lock, _, err := mgr.lockLocations()
+	if err != nil {
+		return err
+	}
+
+	defer lock.Unlock()
+
+	// Check if a location with this name already exists
+	kvp, _, err := mgr.cc.KV().Get(path.Join(consulutil.LocationsPrefix, configuration.Name), nil)
+	if err != nil {
+		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
+	}
+	if kvp == nil {
+		return errors.WithStack(locationNotFoundError{})
 	}
 
 	return mgr.setLocationConfiguration(configuration)
@@ -350,11 +401,33 @@ func (mgr *locationManager) setLocationConfiguration(configuration LocationConfi
 	return err
 }
 
-func (mgr *locationManager) removeLocation(locationName, locationType string) error {
-
-	_, err := mgr.cc.KV().Delete(path.Join(consulutil.LocationsPrefix, locationName), nil)
+func (mgr *locationManager) removeLocation(locationName string) error {
+	// Check if a location with this name already exists
+	kvp, _, err := mgr.cc.KV().Get(path.Join(consulutil.LocationsPrefix, locationName), nil)
+	if err != nil {
+		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
+	}
+	if kvp == nil {
+		// locationName location does not exist
+		return errors.WithStack(locationNotFoundError{})
+	}
+	//It exits, delete it.
+	_, err = mgr.cc.KV().Delete(path.Join(consulutil.LocationsPrefix, locationName), nil)
 	if err != nil {
 		return errors.Wrapf(err, "failed to remove location %s", locationName)
 	}
 	return err
+}
+
+// Return true only if a hostpool location with name locationName exists
+func (mgr *locationManager) checkIsHotPoolLocationType(locationName string) (bool, error) {
+	hostpoolLocations, err := mgr.hpAdapter.ListLocations()
+	exists := false
+	for _, location := range hostpoolLocations {
+		if location == locationName {
+			exists = true
+			break
+		}
+	}
+	return exists, err
 }